move ExternalLib
