+++ designer/client/src/components/toolbarComponents/DefaultToolbarPanel.tsx
@@ -1,6 +1,6 @@
 import React, { PropsWithChildren, ReactElement, useMemo } from "react";
 import { useTranslation } from "react-i18next";
-import { splitUrl } from "../../containers/ExternalLib";
+import { splitUrl } from "../../ExternalLib";
 import { ToolbarButtons } from "./toolbarButtons";
 import { ToolbarConfig } from "../toolbarSettings/types";
 import { ToolbarWrapper } from "./toolbarWrapper/ToolbarWrapper";
